app/requirements.txt
fastapi==0.116.1
httpx==0.28.1
jinja2==3.1.6
pydantic-settings==2.10.1
PyJWT==2.10.1
uvicorn[standard]==0.35.0
